+++ CHANGELOG.md
@@ -15,6 +15,7 @@ CHANGELOG
 - retain schema diagram custom layout on schema change
 - support index and trigger keywords in meta search
 - omit comments from JSON export
+- underline matched texts in meta search results
 - fix loading user-specific configuration file
 - fix schema diagram not loading entity positions from last config
 - fix schema diagram adopting custom layout on clicking an entity