+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 CHANGELOG
 =========
 
-2.1, 2022-09-12
+2.1, 2022-11-01
 ---------------
 - add button panels to data and schema tabs
 - add column and key toggles to schema diagram
@@ -16,6 +16,7 @@ CHANGELOG
 - support index and trigger keywords in meta search
 - omit comments from JSON export
 - underline matched texts in meta search results
+- include journal file sizes in database size
 - fix loading user-specific configuration file
 - fix schema diagram not loading entity positions from last config
 - fix schema diagram adopting custom layout on clicking an entity